+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Consequently, all dependent packages are upgraded as well:
 - Upgrade cernml-coi-funcs requirement to 0.2.1.
 - Upgrade cern-awake-env requirement to 0.15.0.
 - Upgrade cern-leir-transfer-line-env requirement to 0.2.0.
-- Upgrade cern-sps-tune-env requirement to 0.2.0.
+- Upgrade cern-sps-tune-env requirement to 0.2.1.
 - Upgrade cern-sps-zs-alignment-env requirement to 0.2.0.
 
 ## v0.1.3